# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Movie Web App API
  description: |
    REST API for Movie Web App.

    ## Authentication
    Using **Google OAuth2** for login and **JWT Bearer Token** for accessing endpoints that require authentication.

    ## Roles
    - **guest** — Public access without authentication
    - **user** — User that has logged in via Google OAuth
    - **admin** — Administrator with full access
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://movie-ku-api.vercel.app
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from Google OAuth login
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
      required: [success, message]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
      required: [success, message]

    # --- Auth ---
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Access token expiration time in seconds
          example: 300
      required: [access_token, refresh_token, token_type, expires_in]
    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
        profile_picture:
          type: string
          nullable: true
        oauth_provider:
          type: string
          nullable: true
      required: [id, name, email, role]

    # --- Genre ---
...
paths:
...
  /api/auth/google/callback:
    get:
      tags: [Auth]
      summary: Google OAuth2 callback
      operationId: googleCallback
      description: Handles the OAuth2 callback from Google. Exchanges auth code for tokens, creates/updates user, returns JWT tokens.
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      description: Exchange a valid refresh token for a new access token pair.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      operationId: logout
      description: |
        Logout endpoint. Since JWT is stateless, actual token invalidation
        requires a token blacklist (e.g. Redis). Client should discard tokens.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== MOVIES (PUBLIC) ====================
...